fix: anoa_e2e_as_to_device (#4)
* Add experimental config option to send to-device messages to AS's

* Add a new ephemeral AS handler for to_device message edus

Here we add the ability for the application service ephemeral message
processor to handle new events on the "to_device" stream.

We keep track of a stream id (token) per application service, and every
time a new to-device message comes in, for each appservice we pull the
messages between the last-recorded and current stream id and check
whether any of the messages are for a user in that appservice's user
namespace.

get_new_messages is implemented in the next commit.

since we rebased off latest develop.

* Allow setting/getting stream id per appservice for to-device messages

* Add database method to fetch to-device messages by user_ids from db

This method is quite similar to the one below, except that it doesn't
support device ids, and supports querying with more than one user id,
both of which are relevant to application services.

The results are also formatted in a different data structure, so I'm
not sure how much we could really share here between the two methods.

* Add a to_device_stream_id column to the application_services_state table

This is for tracking the stream id that each application service has
been sent up to. In other words, there shouldn't be any need to process
stream ids below the recorded one here as the AS should have already
received them.

Note that there is no reliability built-in here. Reliability of delivery
if intended for a separate PR.

* Add tests

I decided to spin up another test class for this as the existing one is
1. quite old and 2. was mocking away too much of the infrastructure to
my liking. I've named the new class alluding to ephemeral messages, and
while we already have some ephemeral tests in AppServiceHandlerTestCase,
ideally I'd like to migrate those over.

There's two new tests here. One for testing that to-device messages for
a local user are received by any application services that have
registered interest in that user - and that those that haven't won't
receive those messages.

The next test is similar, but tests with a whole bunch of to-device
messages. Rather than actually registering tons of devices - which would
make for a very slow unit test - we just directly insert them into the
database.

* Changelog

* Make msc2409_to_device_messages_enabled private; remove unnecessary check

The second check for self._msc2409_to_device_messages_enabled was not necessary. It's
already checked in notify_interested_services_ephemeral earlier.

* Move stream filter back into AppserviceHandler

* Broaden type hints; update comment

* Deduplicate ephemeral events to send conditional

Test cases needed to be updated, as we now always call
submit_ephemeral_events_for_as, it may just be with an
empty events list.

* _handle_to_device -> _get_to_device_messages

* Rename ApplicationServiceEphemeralEventsTestCase

* Rename user1, user2 in tests to something more useful

* Simplify registration of appservices in tests

* Fix existing unit tests

There is so much mocking going on here. I look forward to replacing these one day.

* rename set_type_stream_id_for_appservice -> set_appservice_stream_type_pos

* Add some FIXME comments

* Add comment on why we don't NOT NULL to_device_stream_id

* Refactor and generalise the sending of arbitrary fields over AS transactions

Things were starting to get a little inflexible as we kept adding new
types of data to send to application services. It's better to just
have one method for adding data to AS transactions, than one for
each type of data.

Note that subsequent PRs will need to add device lists, one-time keys
and fallback keys to these transactions. Adding those are additional
arguments to a method is much nicer than a new method for each one.

Plus with this setup we can add multiple different types of data at
once without kicking off an AS transaction for each type. This will
be useful for OTK/fallback keys, as we plan to lazily attach those
when handling other event types.

* Fix tests to mock _TransactionController.send of ApplicationServiceScheduler.enqueue*

With enqueue_for_appservice being called per-event per-appservice, it's
not a great target for mocking. So we use _TransactionController.send
instead, to track things that are actually sent out to AS's.

This also has the benefit of testing a wider bit of the AS txn pipeline

* Add to-device messages as their own special section in AS txns

* Insert to-device messages into the new to-device part of AS txns

* Modify tests to handle new location of to-device messages in AS txns

* Fix calls to create_appservice_txn in tests

* Update synapse/storage/schema/main/delta/65/06_msc2409_add_device_id_appservice_stream_type.sql

Co-authored-by: Richard van der Hoff <1389908+richvdh@users.noreply.github.com>

* rename recipient_user_id_device_id_to_messages -> recipient_device_to_messages

* lint

* Refactor storage methods to retrieve to-device messages

This commit refactors the previously rather duplicated 'get_new_messages_for_device' and
'get_new_messages' methods into one new private method with combined logic, and two small
public methods. The public methods expose the correct interface for querying to-device
messages for either a single device (where a limit can be used) and multiple devices
(where using a limit is infeasible).

* Update old references to get_new_messages_for_device

Co-authored-by: Andrew Morgan <andrew@amorgan.xyz>
Co-authored-by: Andrew Morgan <1342360+anoadragon453@users.noreply.github.com>
Co-authored-by: Richard van der Hoff <1389908+richvdh@users.noreply.github.com>